server: add CE stub page at server/src/app/msp/extensions/[id]/page.tsx and remove misplaced EE copy
